add confirmation prompts on deploy and remove

+++ index.js
@@ -4,6 +4,7 @@ const path = require('path');
 
 const _ = require('lodash');
 const BbPromise = require('bluebird');
+const Confirm = require('prompt-confirm');
 
 const bucketUtils = require('./lib/bucketUtils');
 const configure = require('./lib/configure');
@@ -59,36 +60,34 @@ class Client {
   _removeDeployedResources() {
     this._validateConfig();
 
-    const safetyDelay = 3000;
     const bucketName = this.options.bucketName;
 
-    this.serverless.cli.log(
-      `Preparing to empty and remove bucket '${bucketName}', waiting for ${safetyDelay /
-        1000} seconds...`
-    );
-
-    return BbPromise.delay(safetyDelay)
-      .then(() => {
-        this.serverless.cli.log(`Looking for bucket...`);
-        return bucketUtils.bucketExists(this.aws, bucketName);
-      })
-      .then(exists => {
-        if (exists) {
-          this.serverless.cli.log(`Deleting all objects from bucket...`);
-          return bucketUtils
-            .emptyBucket(this.aws, bucketName)
-            .then(() => {
-              this.serverless.cli.log(`Removing bucket...`);
-              return bucketUtils.deleteBucket(this.aws, bucketName);
-            })
-            .then(() => {
-              this.serverless.cli.log(
-                `Success! Your files have been removed and your bucket has been deleted`
-              );
-            });
-        } else {
-          this.serverless.cli.log(`Bucket does not exist`);
+    return new Confirm(`Are you sure you want to delete bucket '${bucketName}'?`)
+      .run()
+      .then(goOn => {
+        if (goOn) {
+          this.serverless.cli.log(`Looking for bucket...`);
+          return bucketUtils.bucketExists(this.aws, bucketName).then(exists => {
+            if (exists) {
+              this.serverless.cli.log(`Deleting all objects from bucket...`);
+              return bucketUtils
+                .emptyBucket(this.aws, bucketName)
+                .then(() => {
+                  this.serverless.cli.log(`Removing bucket...`);
+                  return bucketUtils.deleteBucket(this.aws, bucketName);
+                })
+                .then(() => {
+                  this.serverless.cli.log(
+                    `Success! Your files have been removed and your bucket has been deleted`
+                  );
+                });
+            } else {
+              this.serverless.cli.log(`Bucket does not exist`);
+            }
+          });
         }
+        this.serverless.cli.log('Bucket not removed');
+        return BbPromise.resolve();
       })
       .catch(error => {
         return BbPromise.reject(new this.error(error));
@@ -117,65 +116,93 @@ class Client {
     const redirectAllRequestsTo = this.options.redirectAllRequestsTo || null;
     const routingRules = this.options.routingRules || null;
 
-    this.serverless.cli.log(`Deploying client to bucket '${bucketName}' in region '${region}'...`);
+    const deployDescribe = ['This deployment will:'];
 
-    this.serverless.cli.log(`Looking for bucket...`);
-    return bucketUtils
-      .bucketExists(this.aws, bucketName)
-      .then(exists => {
-        if (exists) {
-          this.serverless.cli.log(`Bucket found...`);
-          if (this.cliOptions['delete-contents'] === false) {
-            this.serverless.cli.log(`Keeping current bucket contents...`);
-            return BbPromise.resolve();
-          }
+    if (this.cliOptions['delete-contents']) {
+      deployDescribe.push(`- Remove all existing files from bucket '${bucketName}'`);
+    }
+    deployDescribe.push(
+      `- Upload all files from '${distributionFolder}' to bucket '${bucketName}'`
+    );
+    if (this.cliOptions['config-change'] !== false) {
+      deployDescribe.push(`- Set (and overwrite) bucket '${bucketName}' configuration`);
+    }
+    if (this.cliOptions['policy-change'] !== false) {
+      deployDescribe.push(`- Set (and overwrite) bucket '${bucketName}' bucket policy`);
+    }
+    if (this.cliOptions['cors-change'] !== false) {
+      deployDescribe.push(`- Set (and overwrite) bucket '${bucketName}' CORS policy`);
+    }
 
-          this.serverless.cli.log(`Deleting all objects from bucket...`);
-          return bucketUtils.emptyBucket(this.aws, bucketName);
-        } else {
-          this.serverless.cli.log(`Bucket does not exist. Creating bucket...`);
-          return bucketUtils.createBucket(this.aws, bucketName);
+    deployDescribe.forEach(m => this.serverless.cli.log(m));
+
+    return new Confirm(`Do you want to proceed?`)
+      .run()
+      .then(goOn => {
+        if (goOn) {
+          this.serverless.cli.log(`Looking for bucket...`);
+          return bucketUtils
+            .bucketExists(this.aws, bucketName)
+            .then(exists => {
+              if (exists) {
+                this.serverless.cli.log(`Bucket found...`);
+                if (this.cliOptions['delete-contents'] === false) {
+                  this.serverless.cli.log(`Keeping current bucket contents...`);
+                  return BbPromise.resolve();
+                }
+
+                this.serverless.cli.log(`Deleting all objects from bucket...`);
+                return bucketUtils.emptyBucket(this.aws, bucketName);
+              } else {
+                this.serverless.cli.log(`Bucket does not exist. Creating bucket...`);
+                return bucketUtils.createBucket(this.aws, bucketName);
+              }
+            })
+            .then(() => {
+              if (this.cliOptions['config-change'] === false) {
+                this.serverless.cli.log(`Retaining existing bucket configuration...`);
+                return BbPromise.resolve();
+              }
+              this.serverless.cli.log(`Configuring bucket...`);
+              return configure.configureBucket(
+                this.aws,
+                bucketName,
+                indexDoc,
+                errorDoc,
+                redirectAllRequestsTo,
+                routingRules
+              );
+            })
+            .then(() => {
+              if (this.cliOptions['policy-change'] === false) {
+                this.serverless.cli.log(`Retaining existing bucket policy...`);
+                return BbPromise.resolve();
+              }
+              this.serverless.cli.log(`Configuring policy for bucket...`);
+              return configure.configurePolicyForBucket(this.aws, bucketName);
+            })
+            .then(() => {
+              if (this.cliOptions['cors-change'] === false) {
+                this.serverless.cli.log(`Retaining existing bucket CORS configuration...`);
+                return BbPromise.resolve();
+              }
+              this.serverless.cli.log(`Configuring CORS for bucket...`);
+              return configure.configureCorsForBucket(this.aws, bucketName);
+            })
+            .then(() => {
+              this.serverless.cli.log(`Uploading client files to bucket...`);
+              return uploadDirectory(this.aws, bucketName, clientPath, headerSpec);
+            })
+            .then(() => {
+              this.serverless.cli.log(
+                `Success! Your site should be available at http://${bucketName}.${
+                  regionUrls[region]
+                }/`
+              );
+            });
         }
-      })
-      .then(() => {
-        if (this.cliOptions['config-change'] === false) {
-          this.serverless.cli.log(`Retaining existing bucket configuration...`);
-          return BbPromise.resolve();
-        }
-        this.serverless.cli.log(`Configuring bucket...`);
-        return configure.configureBucket(
-          this.aws,
-          bucketName,
-          indexDoc,
-          errorDoc,
-          redirectAllRequestsTo,
-          routingRules
-        );
-      })
-      .then(() => {
-        if (this.cliOptions['policy-change'] === false) {
-          this.serverless.cli.log(`Retaining existing bucket policy...`);
-          return BbPromise.resolve();
-        }
-        this.serverless.cli.log(`Configuring policy for bucket...`);
-        return configure.configurePolicyForBucket(this.aws, bucketName);
-      })
-      .then(() => {
-        if (this.cliOptions['cors-change'] === false) {
-          this.serverless.cli.log(`Retaining existing bucket CORS configuration...`);
-          return BbPromise.resolve();
-        }
-        this.serverless.cli.log(`Configuring CORS for bucket...`);
-        return configure.configureCorsForBucket(this.aws, bucketName);
-      })
-      .then(() => {
-        this.serverless.cli.log(`Uploading client files to bucket...`);
-        return uploadDirectory(this.aws, bucketName, clientPath, headerSpec);
-      })
-      .then(() => {
-        this.serverless.cli.log(
-          `Success! Your site should be available at http://${bucketName}.${regionUrls[region]}/`
-        );
+        this.serverless.cli.log('Deployment cancelled');
+        return BbPromise.resolve();
       })
       .catch(error => {
         return BbPromise.reject(new this.error(error));
